requirements.txt
Flask==3.0.2
pypandoc_binary==1.14
pdfkit==1.0.0
